refactor out LayoutGrid class
mostly just pushing code around here, but should be able to add more
methods now
